Switching back to bind mount.
Tested volume mount, and Codespaces still said “insecure operation.” I’d assuming it was the mount since it mentioned the mount. I am commenting out the mount feature altogether to see if it is the mount line. I will inspect the default value situation to see if that feature got implemented and the /dev/null mounts are the cause. For now, I’m just seeing if it’s the mount line. If so, I’ll make another commit.